+++ pelican/tests/test_cache.py
@@ -6,7 +6,7 @@ from shutil import rmtree
 from tempfile import mkdtemp
 
 from pelican.generators import ArticlesGenerator, PagesGenerator
-from pelican.tests.support import get_settings, unittest
+from pelican.tests.support import get_context, get_settings, unittest
 
 try:
     from unittest.mock import MagicMock
@@ -29,7 +29,7 @@ class TestCache(unittest.TestCase):
         rmtree(self.temp_cache)
 
     def _get_cache_enabled_settings(self):
-        settings = get_settings(filenames={})
+        settings = get_settings()
         settings['CACHE_CONTENT'] = True
         settings['LOAD_CONTENT_CACHE'] = True
         settings['CACHE_PATH'] = self.temp_cache
@@ -42,20 +42,21 @@ class TestCache(unittest.TestCase):
         settings['PAGE_PATHS'] = ['TestPages']
         settings['DEFAULT_DATE'] = (1970, 1, 1)
         settings['READERS'] = {'asc': None}
+        context = get_context(settings)
 
         def sorted_titles(items):
             return sorted(item.title for item in items)
 
         # Articles
         generator = ArticlesGenerator(
-            context=settings.copy(), settings=settings,
+            context=context.copy(), settings=settings,
             path=CONTENT_DIR, theme=settings['THEME'], output_path=None)
         generator.generate_context()
         uncached_articles = sorted_titles(generator.articles)
         uncached_drafts = sorted_titles(generator.drafts)
 
         generator = ArticlesGenerator(
-            context=settings.copy(), settings=settings,
+            context=context.copy(), settings=settings,
             path=CONTENT_DIR, theme=settings['THEME'], output_path=None)
         generator.generate_context()
         cached_articles = sorted_titles(generator.articles)
@@ -66,7 +67,7 @@ class TestCache(unittest.TestCase):
 
         # Pages
         generator = PagesGenerator(
-            context=settings.copy(), settings=settings,
+            context=context.copy(), settings=settings,
             path=CUR_DIR, theme=settings['THEME'], output_path=None)
         generator.generate_context()
         uncached_pages = sorted_titles(generator.pages)
@@ -74,7 +75,7 @@ class TestCache(unittest.TestCase):
         uncached_draft_pages = sorted_titles(generator.draft_pages)
 
         generator = PagesGenerator(
-            context=settings.copy(), settings=settings,
+            context=context.copy(), settings=settings,
             path=CUR_DIR, theme=settings['THEME'], output_path=None)
         generator.generate_context()
         cached_pages = sorted_titles(generator.pages)
@@ -92,20 +93,21 @@ class TestCache(unittest.TestCase):
         settings['PAGE_PATHS'] = ['TestPages']
         settings['DEFAULT_DATE'] = (1970, 1, 1)
         settings['READERS'] = {'asc': None}
+        context = get_context(settings)
 
         def sorted_titles(items):
             return sorted(item.title for item in items)
 
         # Articles
         generator = ArticlesGenerator(
-            context=settings.copy(), settings=settings,
+            context=context.copy(), settings=settings,
             path=CONTENT_DIR, theme=settings['THEME'], output_path=None)
         generator.generate_context()
         uncached_articles = sorted_titles(generator.articles)
         uncached_drafts = sorted_titles(generator.drafts)
 
         generator = ArticlesGenerator(
-            context=settings.copy(), settings=settings,
+            context=context.copy(), settings=settings,
             path=CONTENT_DIR, theme=settings['THEME'], output_path=None)
         generator.generate_context()
         cached_articles = sorted_titles(generator.articles)
@@ -116,14 +118,14 @@ class TestCache(unittest.TestCase):
 
         # Pages
         generator = PagesGenerator(
-            context=settings.copy(), settings=settings,
+            context=context.copy(), settings=settings,
             path=CUR_DIR, theme=settings['THEME'], output_path=None)
         generator.generate_context()
         uncached_pages = sorted_titles(generator.pages)
         uncached_hidden_pages = sorted_titles(generator.hidden_pages)
 
         generator = PagesGenerator(
-            context=settings.copy(), settings=settings,
+            context=context.copy(), settings=settings,
             path=CUR_DIR, theme=settings['THEME'], output_path=None)
         generator.generate_context()
         cached_pages = sorted_titles(generator.pages)
@@ -139,15 +141,16 @@ class TestCache(unittest.TestCase):
         settings['CONTENT_CACHING_LAYER'] = 'generator'
         settings['DEFAULT_DATE'] = (1970, 1, 1)
         settings['READERS'] = {'asc': None}
+        context = get_context(settings)
 
         generator = ArticlesGenerator(
-            context=settings.copy(), settings=settings,
+            context=context.copy(), settings=settings,
             path=CONTENT_DIR, theme=settings['THEME'], output_path=None)
         generator.generate_context()
         self.assertTrue(hasattr(generator, '_cache'))
 
         generator = ArticlesGenerator(
-            context=settings.copy(), settings=settings,
+            context=context.copy(), settings=settings,
             path=CONTENT_DIR, theme=settings['THEME'], output_path=None)
         generator.readers.read_file = MagicMock()
         generator.generate_context()
@@ -167,15 +170,16 @@ class TestCache(unittest.TestCase):
         """Test raw article content caching at the reader level"""
         settings = self._get_cache_enabled_settings()
         settings['READERS'] = {'asc': None}
+        context = get_context(settings)
 
         generator = ArticlesGenerator(
-            context=settings.copy(), settings=settings,
+            context=context.copy(), settings=settings,
             path=CONTENT_DIR, theme=settings['THEME'], output_path=None)
         generator.generate_context()
         self.assertTrue(hasattr(generator.readers, '_cache'))
 
         generator = ArticlesGenerator(
-            context=settings.copy(), settings=settings,
+            context=context.copy(), settings=settings,
             path=CONTENT_DIR, theme=settings['THEME'], output_path=None)
         readers = generator.readers.readers
         for reader in readers.values():
@@ -191,9 +195,10 @@ class TestCache(unittest.TestCase):
         used in --ignore-cache or autoreload mode"""
         settings = self._get_cache_enabled_settings()
         settings['READERS'] = {'asc': None}
+        context = get_context(settings)
 
         generator = ArticlesGenerator(
-            context=settings.copy(), settings=settings,
+            context=context.copy(), settings=settings,
             path=CONTENT_DIR, theme=settings['THEME'], output_path=None)
         generator.readers.read_file = MagicMock()
         generator.generate_context()
@@ -202,7 +207,7 @@ class TestCache(unittest.TestCase):
 
         settings['LOAD_CONTENT_CACHE'] = False
         generator = ArticlesGenerator(
-            context=settings.copy(), settings=settings,
+            context=context.copy(), settings=settings,
             path=CONTENT_DIR, theme=settings['THEME'], output_path=None)
         generator.readers.read_file = MagicMock()
         generator.generate_context()
@@ -217,15 +222,16 @@ class TestCache(unittest.TestCase):
         settings['CONTENT_CACHING_LAYER'] = 'generator'
         settings['PAGE_PATHS'] = ['TestPages']
         settings['READERS'] = {'asc': None}
+        context = get_context(settings)
 
         generator = PagesGenerator(
-            context=settings.copy(), settings=settings,
+            context=context.copy(), settings=settings,
             path=CUR_DIR, theme=settings['THEME'], output_path=None)
         generator.generate_context()
         self.assertTrue(hasattr(generator, '_cache'))
 
         generator = PagesGenerator(
-            context=settings.copy(), settings=settings,
+            context=context.copy(), settings=settings,
             path=CUR_DIR, theme=settings['THEME'], output_path=None)
         generator.readers.read_file = MagicMock()
         generator.generate_context()
@@ -241,15 +247,16 @@ class TestCache(unittest.TestCase):
         settings = self._get_cache_enabled_settings()
         settings['PAGE_PATHS'] = ['TestPages']
         settings['READERS'] = {'asc': None}
+        context = get_context(settings)
 
         generator = PagesGenerator(
-            context=settings.copy(), settings=settings,
+            context=context.copy(), settings=settings,
             path=CUR_DIR, theme=settings['THEME'], output_path=None)
         generator.generate_context()
         self.assertTrue(hasattr(generator.readers, '_cache'))
 
         generator = PagesGenerator(
-            context=settings.copy(), settings=settings,
+            context=context.copy(), settings=settings,
             path=CUR_DIR, theme=settings['THEME'], output_path=None)
         readers = generator.readers.readers
         for reader in readers.values():
@@ -266,9 +273,10 @@ class TestCache(unittest.TestCase):
         settings = self._get_cache_enabled_settings()
         settings['PAGE_PATHS'] = ['TestPages']
         settings['READERS'] = {'asc': None}
+        context = get_context(settings)
 
         generator = PagesGenerator(
-            context=settings.copy(), settings=settings,
+            context=context.copy(), settings=settings,
             path=CUR_DIR, theme=settings['THEME'], output_path=None)
         generator.readers.read_file = MagicMock()
         generator.generate_context()
@@ -277,7 +285,7 @@ class TestCache(unittest.TestCase):
 
         settings['LOAD_CONTENT_CACHE'] = False
         generator = PagesGenerator(
-            context=settings.copy(), settings=settings,
+            context=context.copy(), settings=settings,
             path=CUR_DIR, theme=settings['THEME'], output_path=None)
         generator.readers.read_file = MagicMock()
         generator.generate_context()