+++ pelican/tests/test_pelican.py
@@ -112,7 +112,7 @@ class TestPelican(LoggedTestCase):
         self.assertDirsEqual(
             self.temp_path, os.path.join(OUTPUT_PATH, 'basic'))
         self.assertLogCountEqual(
-            count=3,
+            count=1,
             msg="Unable to find.*skipping url replacement",
             level=logging.WARNING)